node_function(int) -> node_function(size_t)

+++ src/NodeBase.h
@@ -30,7 +30,7 @@ namespace Banyan {
   };
 
 
-  typedef NodeReturnStatus (node_function)(int identifier);
+  typedef NodeReturnStatus (node_function)(size_t identifier);
 
 
   class NodeBase {
